Moved different components into separate sub-folders, reduced size of wizard-config.jsx by dividing the data inside into 2 files - wizard_categories.jsx and wizard_onboarding_methods.jsx.
