+++ api-catalog-ui/frontend/src/components/Dashboard/Dashboard.test.jsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 // tslint:disable-next-line:no-implicit-dependencies
 import { shallow } from 'enzyme';
 import Dashboard from './Dashboard';
-import { data } from '../Wizard/wizard_config';
+import { categoryData } from '../Wizard/configs/wizard_categories';
 
 const ajaxError = {
     message: 'ajax Error 404',
@@ -116,7 +116,7 @@ describe('>>> Dashboard component tests', () => {
                 fetchTilesStop={jest.fn()}
                 clearService={jest.fn()}
                 clear={jest.fn()}
-                inputData={data}
+                inputData={categoryData}
             />
         );
         const instance = wrapper.instance();
