+++ api-catalog-ui/frontend/src/components/Wizard/WizardComponents/WizardInputs.test.jsx
