+++ api-catalog-ui/frontend/src/components/Wizard/WizardComponents/WizardNavigation.test.jsx
