+++ api-catalog-ui/frontend/src/components/Wizard/configs/wizard_defaults.jsx
